PRKT-88 Rename Serial Driver
